Correct the restart-shutdown command

+++ zsh/.config/zsh/scripts/restart-shutdown.sh
@@ -1,8 +1,11 @@
 #!/bin/sh
 # Restart or shutdown the system
 
+CURRENT_DIR="$(dirname "$(realpath "$0")")"
+. "$CURRENT_DIR"/helpers.sh
+
 restart() {
-    if [ -d "/etc/systemd/" ] ; then
+    if [ -d "/etc/systemd/" ] && [ -d "/etc/apt" ] ; then
         sudo systemctl reboot
     else
         sudo reboot
@@ -10,8 +13,8 @@ restart() {
 }
 
 shutdown() {
-    if [ -d "/etc/systemd/" ] ; then
-        systemctl poweroff
+    if [ -d "/etc/systemd/" ] && [ -d "/etc/apt" ] ; then
+        sudo systemctl poweroff
     else
         sudo poweroff
     fi
